requirements.txt
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
tqdm>=4.65.0
python-dotenv>=1.0.0 